+++ golang/kurtosis-engine-server-consts/engine_server_consts.go
@@ -1,0 +1,9 @@
+package kurtosis_engine_server_consts
+
+const (
+	// Protocol that the engine server container listens on
+	ListenProtocol = "tcp"
+
+	// The port that the engine server container listens on
+	ListenPort = 7443
+)