+++ typescript/src/kurtosis_engine_rpc_api_bindings/engine_service_grpc_pb.d.ts
@@ -4,16 +4,21 @@
 // file: engine_service.proto
 
 import * as engine_service_pb from "./engine_service_pb";
+import * as google_protobuf_empty_pb from "google-protobuf/google/protobuf/empty_pb";
 import * as grpc from "grpc";
 
 interface IEngineServiceService extends grpc.ServiceDefinition<grpc.UntypedServiceImplementation> {
   createEnclave: grpc.MethodDefinition<engine_service_pb.CreateEnclaveArgs, engine_service_pb.CreateEnclaveResponse>;
+  getEnclave: grpc.MethodDefinition<engine_service_pb.GetEnclaveArgs, engine_service_pb.GetEnclaveResponse>;
+  destroyEnclave: grpc.MethodDefinition<engine_service_pb.DestroyEnclaveArgs, google_protobuf_empty_pb.Empty>;
 }
 
 export const EngineServiceService: IEngineServiceService;
 
 export interface IEngineServiceServer extends grpc.UntypedServiceImplementation {
   createEnclave: grpc.handleUnaryCall<engine_service_pb.CreateEnclaveArgs, engine_service_pb.CreateEnclaveResponse>;
+  getEnclave: grpc.handleUnaryCall<engine_service_pb.GetEnclaveArgs, engine_service_pb.GetEnclaveResponse>;
+  destroyEnclave: grpc.handleUnaryCall<engine_service_pb.DestroyEnclaveArgs, google_protobuf_empty_pb.Empty>;
 }
 
 export class EngineServiceClient extends grpc.Client {
@@ -21,4 +26,10 @@ export class EngineServiceClient extends grpc.Client {
   createEnclave(argument: engine_service_pb.CreateEnclaveArgs, callback: grpc.requestCallback<engine_service_pb.CreateEnclaveResponse>): grpc.ClientUnaryCall;
   createEnclave(argument: engine_service_pb.CreateEnclaveArgs, metadataOrOptions: grpc.Metadata | grpc.CallOptions | null, callback: grpc.requestCallback<engine_service_pb.CreateEnclaveResponse>): grpc.ClientUnaryCall;
   createEnclave(argument: engine_service_pb.CreateEnclaveArgs, metadata: grpc.Metadata | null, options: grpc.CallOptions | null, callback: grpc.requestCallback<engine_service_pb.CreateEnclaveResponse>): grpc.ClientUnaryCall;
+  getEnclave(argument: engine_service_pb.GetEnclaveArgs, callback: grpc.requestCallback<engine_service_pb.GetEnclaveResponse>): grpc.ClientUnaryCall;
+  getEnclave(argument: engine_service_pb.GetEnclaveArgs, metadataOrOptions: grpc.Metadata | grpc.CallOptions | null, callback: grpc.requestCallback<engine_service_pb.GetEnclaveResponse>): grpc.ClientUnaryCall;
+  getEnclave(argument: engine_service_pb.GetEnclaveArgs, metadata: grpc.Metadata | null, options: grpc.CallOptions | null, callback: grpc.requestCallback<engine_service_pb.GetEnclaveResponse>): grpc.ClientUnaryCall;
+  destroyEnclave(argument: engine_service_pb.DestroyEnclaveArgs, callback: grpc.requestCallback<google_protobuf_empty_pb.Empty>): grpc.ClientUnaryCall;
+  destroyEnclave(argument: engine_service_pb.DestroyEnclaveArgs, metadataOrOptions: grpc.Metadata | grpc.CallOptions | null, callback: grpc.requestCallback<google_protobuf_empty_pb.Empty>): grpc.ClientUnaryCall;
+  destroyEnclave(argument: engine_service_pb.DestroyEnclaveArgs, metadata: grpc.Metadata | null, options: grpc.CallOptions | null, callback: grpc.requestCallback<google_protobuf_empty_pb.Empty>): grpc.ClientUnaryCall;
 }