+++ docs/changelog.md
@@ -1,4 +1,11 @@
 # TBD
+### Features
+* Added `engine_server_consts` which contains constants related to the Kurtosis Engine Server (e.g. ListenProtocol)
+* Added `api_container_log_level` argument in `CreateEnclaveArgs` to specified the log level that the API container running in the enclave should have
+
+### Breaking changes
+* Renamed `network_ip` field in `CreateEnclaveResponse` protobuf message to `network_cidr` because the value will contain IP address and Mask
+    * Users should use the new field `network_cidr` instead the old `network_ip` and parse it to get the network ip address and mask
 
 # 0.1.1
 ### Features